\name{Guiana_Culex}
\docType{data}
\alias{Guiana_Culex}
\alias{Thorax_Guiana_Culex}
\alias{Thorax_Guiana_Culex1}
\alias{Thorax_Guiana_Culex2}
\alias{Thorax_Guiana_Culex3}
\alias{Legs_Guiana_Culex}
\alias{Legs_Guiana_Culex1}
\alias{Legs_Guiana_Culex2}
\alias{Legs_Guiana_Culex3}
\title{Guiana_Culex}
\description{

}
\usage{Guiana_Culex}
\format{Six lists of 225 spectra}
\source{Refka}
\references{

}
\keyword{datasets}
