\name{MSProfileR-package}


\alias{MSProfileR-package}
\alias{MSProfileR}

\docType{package}

\title{
  MSProfileR
}

\description{
  ...
}

\details{
  \tabular{ll}{
    Package: \tab  MSProfileR \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.1 \cr
    Date: \tab  13/12/2023 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
  Refka Ben Hamouda  <refkabenhamouda10@gmail.com> \cr
}
\references{

}
\seealso{

}
\examples{

}
